did some pending

+++ database/migrations/2023_11_17_123928_create_attributes_table.php
@@ -14,7 +14,8 @@ return new class extends Migration
         Schema::create('attributes', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->string('name')->unique();
-            $table->integer('attribute_no')->unique();
+            // $table->integer('attribute_no')->unique();
+            $table->decimal('attribute_no', 10, 1)->unique();
             $table->timestamps();
         });
     }
